+++ spec/factories/claims.rb
@@ -25,6 +25,7 @@ FactoryBot.define do
       teacher_reference_number { generate(:teacher_reference_number) }
       national_insurance_number { generate(:national_insurance_number) }
       email_address { generate(:email_address) }
+      bank_or_building_society { :personal_bank_account }
       banking_name { "Jo Bloggs" }
       bank_sort_code { rand(100000..999999) }
       bank_account_number { rand(10000000..99999999) }